New class and stack implimentation
Implemented a new class Game that controls the game structure and movements of player and game-over conditions are implemented in it. Further more, a Stack class is implemented for keeping the control of undo moves and its functions are used in the player class to push each move and upon undo pop move to go back to the previous node.